backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
anthropic
numpy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
